+++ src/cmd/ksh93/include/fcin.h
@@ -33,7 +33,7 @@ typedef struct _fcin
 	unsigned char	*fcptr;		/* pointer to next input char */
 	unsigned char	fcchar;		/* saved character */
 	short		fclen;		/* last multibyte char len */
-	void (*fcfun)(Sfio_t*,const char*,int,void*);	/* advance function */
+	void (*fcfun)(Sfio_t*,const char*,ssize_t,void*);	/* advance function */
 	void		*context;	/* context pointer */
 	int		fcleft;		/* for multibyte boundary */
 	Sfoff_t		fcoff;		/* offset for last read */
@@ -46,9 +46,9 @@ typedef struct _fcin
 #   define fcmbget(x)	(fcget())
 #endif
 #define fcfile()	(_Fcin._fcfile)
-#define fcgetc()	(*_Fcin.fcptr++ ? (int)_Fcin.fcptr[-1] : fcfill())
-#define fcget()		((int)(*_Fcin.fcptr++))
-#define fcpeek(n)	((int)_Fcin.fcptr[n])
+#define fcgetc()	(*_Fcin.fcptr++ ? _Fcin.fcptr[-1] : fcfill())
+#define fcget()		(*_Fcin.fcptr++)
+#define fcpeek(n)	(_Fcin.fcptr[n])
 #define fcseek(n)	((char*)(_Fcin.fcptr+=(n)))
 #define fcfirst()	((char*)_Fcin.fcbuff)
 #define fclast()	((char*)_Fcin.fclast)
@@ -58,9 +58,9 @@ typedef struct _fcin
 #define fcsave(x)	(*(x) = _Fcin)
 #define fcrestore(x)	(_Fcin = *(x))
 extern int		fcfill(void);
-extern int		fcfopen(Sfio_t*);
-extern int		fcclose(void);
-void			fcnotify(void(*)(Sfio_t*,const char*,int,void*),void*);
+extern ssize_t		fcfopen(Sfio_t*);
+extern ssize_t		fcclose(void);
+void			fcnotify(void(*)(Sfio_t*,const char*,ssize_t,void*),void*);
 
 extern Fcin_t		_Fcin;		/* used by macros */
